+++ addons_custom/dz23_whatsapp/models/__init__.py
@@ -1,4 +1,5 @@
 from . import (
+    business_action,
     message_event,
     message_inbox,
     message_outbox,